Multiple-Disease-Prediction-Webapp/Frontend/requirements-grok.txt
groq
python-dotenv>=1.0.1
